# doc/api.yml
openapi: 3.0.3
info:
  title: Cinema API
  version: 1.0.0
  description: Strict OpenAPI specification for the current poja-cinema REST API.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /reservations:
    get:
      tags:
        - Reservations
      summary: List reservations
      operationId: getReservations
      security:
        - bearerAuth: []
      x-required-roles:
        - EMPLOYEE
        - MANAGER
      responses:
        "200":
          description: Reservations returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReservationResponse"
        "401":
          description: Invalid or missing JWT token
        "403":
          description: User is not EMPLOYEE or MANAGER

    post:
      tags:
        - Reservations
      summary: Create reservation for the authenticated client
      description: The reservation owner is the authenticated CLIENT from the JWT token. The request body must not contain userId.
      operationId: createReservation
      security:
        - bearerAuth: []
      x-required-roles:
        - CLIENT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
            example:
              seatId: 11111111-1111-4111-8111-111111111111
              projectionId: 22222222-2222-4222-8222-222222222222
      responses:
        "200":
          description: Reservation created with status PENDING
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or missing JWT token
        "403":
          description: User is not CLIENT
        "404":
          description: Seat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Seat already reserved for this projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservations/{id}:
    get:
      tags:
        - Reservations
      summary: Get reservation by id
      description: CLIENT can read only his own reservation. EMPLOYEE and MANAGER can read any reservation.
      operationId: getReservationById
      security:
        - bearerAuth: []
      x-required-roles:
        - CLIENT
        - EMPLOYEE
        - MANAGER
      parameters:
        - $ref: "#/components/parameters/UuidPathId"
      responses:
        "200":
          description: Reservation found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "401":
          description: Invalid or missing JWT token
        "403":
          description: CLIENT cannot access another user's reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - Reservations
      summary: Update reservation status
      description: CLIENT can update only his own reservation and only to CANCELED. EMPLOYEE and MANAGER can update reservation status.
      operationId: updateReservationStatus
      security:
        - bearerAuth: []
      x-required-roles:
        - CLIENT
        - EMPLOYEE
        - MANAGER
      parameters:
        - $ref: "#/components/parameters/UuidPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReservationStatusRequest"
            examples:
              cancel:
                summary: Cancel reservation
                value:
                  status: CANCELED
              validate:
                summary: Validate reservation
                value:
                  status: SUCCESS
      responses:
        "200":
          description: Reservation status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or missing JWT token
        "403":
          description: User is not allowed to update this reservation with the requested status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidPathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 11111111-1111-4111-8111-111111111111
  schemas:
...
    CreateReservationRequest:
      type: object
      required:
        - seatId
        - projectionId
      additionalProperties: false
      properties:
        seatId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        projectionId:
          type: string
          format: uuid
          example: 22222222-2222-4222-8222-222222222222
    UpdateReservationStatusRequest:
      type: object
      required:
        - status
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/ReservationStatus"
    ReservationResponse:
      type: object
      required:
        - id
        - userId
        - seatId
        - projectionId
        - createdAt
        - status
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          description: Owner user id.
        seatId:
          type: string
          format: uuid
        projectionId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/ReservationStatus"
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      additionalProperties: false
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Resource not found
        path:
          type: string
          example: /reservations/11111111-1111-4111-8111-111111111111
...
    ReservationStatus:
      type: string
      enum:
        - PENDING
        - SUCCESS
        - CANCELED
...